requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
celery>=5.2.7
redis>=4.5.4
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
pydantic>=1.10.7
httpx>=0.24.0
apscheduler>=3.9.1
eventlet>=0.33.3
cryptography>=44.0.2
requests>=2.32.3